requirements.txt
pandas
boto3
requests
pyyaml
xlrd
openpyxl